GB3-1715: Move host listenter to directive
