+++ src/app/map/components/search-window/search-window.component.html
@@ -1,4 +1,4 @@
-<div class="search-window">
+<div class="search-window" searchResultKeyboardNavigation [allSearchResults]="allSearchResults">
   <div class="search-window__searchbar" [ngClass]="{'search-window__searchbar--mobile': screenMode === 'mobile'}">
     <search
       typedTourAnchor="map.search.input"
