+++ src/app/shared/directives/search-result-keyboard-navigation.directive.spec.ts
@@ -1,0 +1,8 @@
+import {SearchResultKeyboardNavigationDirective} from './search-result-keyboard-navigation.directive';
+
+describe('SearchResultKeyboardNavigationDirective', () => {
+  it('should create an instance', () => {
+    const directive = new SearchResultKeyboardNavigationDirective();
+    expect(directive).toBeTruthy();
+  });
+});
